requirements.txt
arrow==0.10.0
astroid==1.6.1
attrs==17.4.0
autopep8==1.3.4
b2==0.7.2
BitcoinExchangeFH==0.2.3
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
coloredlogs==7.1
colorlog==3.1.2
coverage==4.5.1
Distutils==1.0
ebaysdk==2.1.5
flake8==3.5.0
Flask==0.12.2
funcsigs==1.0.2
gcloud==0.18.3
google-api-core==0.1.1
google-auth==1.2.0
google-cloud-core==0.28.0
google-cloud-storage==1.6.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.7.1
httplib2==0.10.3
humanfriendly==4.4
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.9.6
jws==0.1.3
lazy-object-proxy==1.3.1
logfury==0.1.2
lxml==4.1.1
MarkupSafe==1.0
mccabe==0.6.1
munch==2.2.0
nltk==3.2.4
nose==1.3.7
numpy==1.13.1
oauth2client==4.1.2
pluggy==0.6.0
prestapyt==0.8.0
protobuf==3.4.0
py==1.5.2
PyAlgoTrade==0.18
pyasn1==0.3.7
pyasn1-modules==0.1.5
pybars3==0.9.3
pycodestyle==2.3.1
pycryptodome==3.4.3
pyflakes==1.6.0
pylint==1.8.2
PyMeta3==0.5.1
pymongo==3.4.0
PyMySQL==0.7.11
Pyrebase==3.0.27
pytest==3.4.1
pytest-cov==2.5.1
pytest-html==1.16.1
pytest-metadata==1.6.0
python-dateutil==2.6.1
python-jwt==2.0.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.3
qPython==1.2.2
requests==2.18.4
requests-toolbelt==0.7.0
rope==0.10.7
rsa==3.4.2
six==1.11.0
textblob==0.12.0
tqdm==4.15.0
urllib3==1.22
virtualenv==15.1.0
Wand==0.4.4
websocket-client==0.44.0
Werkzeug==0.12.2
wrapt==1.10.11
